backend/shared/pyproject.toml
[project]
name = "agentic-trust-shared"
version = "0.1.0"
description = "Shared utilities for Agentic Trust Platform services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "fastapi>=0.126.0",
    "prometheus-client>=0.21.0",
    "pydantic>=2.10.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "slowapi>=0.1.9",
    "structlog>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["shared/"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

